+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - PR #3029 Update gdf_ numeric types with stdint and move to cudf namespace
 - PR #3070 Move functions.h and related source to legacy
 - PR #2951 Allow set_index to handle a list of column names
+- PR #3067 Java method to return size of device memory buffer
 
 ## Bug Fixes